Improve page transitions
To simplify layout of non-homepage pages, wrap all of the homepage
elements in a .home-page-container, which is absolutely positioned and
goes away when we transition to a non-homepage page.

Making .page divs (inside #content) absolutely positioned isn’t
something I really like. It depends on JS to resize #content when the
browser can do that better. Get rid of that, and just keep the divs in
the flow, making sure to set display:none after fading out.

Clean up a lot of the JS handling page transitions…